move app config service to layout server
